auto create report

+++ main/views.py
@@ -5,11 +5,14 @@ from django.http import JsonResponse
 from handleData.prepare import prepare
 from handleData.analyze import analyze
 import base64, json
+from main.createReport import createReport
 # Create your views here.
 wxb_name = "./handleData/data/wang.xlsx"
 world_name = "./handleData/data/owd.csv"
 export_path = "./main/static/export"
 report_path = "./main/static/report"
+
+
 def index(request):
     return render(request, "index.html")
 
@@ -45,24 +48,25 @@ def saveImage(request):
     if request.method == "POST":
         try:
             baseimg = json.loads(request.POST.get("baseimg"))
-            text = request.POST.get("text")
-            print(text)
-            # with open("%s/%s.png" % (report_path, '2'), 'wb') as f:
-            #     f.write(base64.b64decode(baseimg[21:]))
-            # for key, value in baseimg.items():
-            #     with open("%s/%s.png" % (report_path, key),'wb') as f:
-            #         f.write(base64.b64decode(value[21:]))
-            path = "%s/text.txt" % (report_path)
-            if (os.path.exists(path)):
-                os.remove(path)
-            for line in text.split('\n'):
-                line = line.strip()
-                if len(line) > 3:
-                    with open(path, 'a', encoding='utf-8') as f:
-                        f.write(line+"\n")
+            text = request.POST.get("text","")
 
+            for key, value in baseimg.items():
+                with open("%s/%s.png" % (report_path, key),'wb') as f:
+                    f.write(base64.b64decode(value[21:]))
+
+            if text:
+                path = "%s/text.txt" % (report_path)
+                if (os.path.exists(path)):
+                    os.remove(path)
+                for line in text.split('\n'):
+                    line = line.strip()
+                    if len(line) > 3:
+                        with open(path, 'a', encoding='utf-8') as f:
+                            f.write(line+"\n")
+            createReport()
         except Exception as e:
             return JsonResponse({"error": "Error!\n"+repr(e)})
+
         return JsonResponse({"result": True})
 
     return JsonResponse({"result": "404"})
